refactor(eval): split correlation_engine.rs into module directory
Modularize the 4395-line correlation_engine.rs into focused submodules
for maintainability. No logic, behavior, or API changes -- pure
restructuring with identical test results (366 tests pass).

New structure:
  correlation_engine/
    mod.rs     -- engine struct, impl, timestamp helpers
    types.rs   -- config enums, result types, snapshot
    tests.rs   -- all unit tests
